refactor header navigation
